moving, need to terminate the conditionstions to false
need to terminate the conditions to false. and make sure that the encoders dont resent, and probs need to fix that 1000 thing on controller. need to neck if it gets the next index values, the calculation should be continues from the previous ones

if this doesnt work, we have to use a method to increment the angle in a loop or something